+++ src/vs/workbench/contrib/terminalContrib/chatAgentTools/browser/task/runTaskTool.ts
@@ -12,11 +12,12 @@ import { ILanguageModelsService } from '../../../../chat/common/languageModels.j
 import { CountTokensCallback, IPreparedToolInvocation, IToolData, IToolImpl, IToolInvocation, IToolInvocationPreparationContext, IToolResult, ToolDataSource, ToolProgress } from '../../../../chat/common/languageModelToolsService.js';
 import { ITaskService, ITaskSummary, Task } from '../../../../tasks/common/taskService.js';
 import { ITerminalService } from '../../../../terminal/browser/terminal.js';
-import { getExpectedUserInputKind, pollForOutputAndIdle, promptForMorePolling, racePollingOrPrompt, handleYesNoUserPrompt } from '../bufferOutputPolling.js';
+import { getExpectedUserInputKind, pollForOutputAndIdle, racePollingOrPrompt, handleYesNoUserPrompt } from '../bufferOutputPolling.js';
 import { getOutput } from '../outputHelpers.js';
 import { getTaskDefinition, getTaskForTool } from './taskHelpers.js';
 import { MarkdownString } from '../../../../../../base/common/htmlContent.js';
 import { IConfigurationService } from '../../../../../../platform/configuration/common/configuration.js';
+import { promptForMorePolling } from '../../../../elicitation/browser/elicitation.js';
 
 type RunTaskToolClassification = {
 	taskId: { classification: 'SystemMetaData'; purpose: 'FeatureInsight'; comment: 'The ID of the task.' };
